+++ src/DonorProfiles/Pipeline/Stages/UpdateDonorAvatar.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Give\DonorProfiles\Pipeline\Stages;
+
+class UpdateDonorAvatar {
+
+	protected $data;
+	protected $donor;
+
+	public function __invoke( $payload ) {
+
+		$this->data  = $payload['data'];
+		$this->donor = $payload['donor'];
+
+		$this->updateAvatarInMetaDB();
+
+		return $payload;
+
+	}
+
+	protected function updateAvatarInMetaDB() {
+		$attributeMetaMap = [
+			'avatarId' => '_give_donor_avatar_id',
+		];
+
+		foreach ( $attributeMetaMap as $attribute => $metaKey ) {
+			if ( property_exists( $this->data, $attribute ) ) {
+				$this->donor->update_meta( $metaKey, $this->data->{$attribute} );
+			}
+		}
+	}
+}