+++ src/DonorProfiles/resources/js/app/components/text-control/index.js
@@ -1,18 +1,20 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { toKebabCase } from '../../utils';
+import { toUniqueId } from '../../utils';
 
 import './style.scss';
 
 const TextControl = ( { label, value, onChange, icon, type } ) => {
+	const id = toUniqueId( label );
+
 	return (
 		<div className="give-donor-profile-text-control">
-			{ label && ( <label className="give-donor-profile-text-control__label" htmlFor={ toKebabCase( label ) }>{ label }</label> ) }
+			{ label && ( <label className="give-donor-profile-text-control__label" htmlFor={ id }>{ label }</label> ) }
 			<div className="give-donor-profile-text-control__input">
 				{ icon && (
 					<FontAwesomeIcon icon={ icon } />
 				) }
 				<input
-					id={ label && toKebabCase( label ) }
+					id={ id }
 					type={ type }
 					value={ value }
 					onChange={ ( evt ) => onChange( evt.target.value ) }