+++ src/DonorProfiles/resources/js/app/store/actions.js
@@ -6,3 +6,30 @@ export const addTab = ( tab ) => {
 		},
 	};
 };
+
+export const setProfile = ( profile ) => {
+	return {
+		type: 'SET_PROFILE',
+		payload: {
+			profile,
+		},
+	};
+};
+
+export const setStates = ( states ) => {
+	return {
+		type: 'SET_STATES',
+		payload: {
+			states,
+		},
+	};
+};
+
+export const setFetchingStates = ( fetchingStates ) => {
+	return {
+		type: 'SET_FETCHING_STATES',
+		payload: {
+			fetchingStates,
+		},
+	};
+};