+++ src/DonorProfiles/resources/js/app/store/initialState.js
@@ -1,3 +1,10 @@
+import { getWindowData } from '../utils';
+
 export const initialState = {
 	tabs: {},
+	profile: getWindowData( 'profile' ),
+	id: getWindowData( 'id' ),
+	countries: getWindowData( 'countries' ),
+	states: getWindowData( 'states' ),
+	fetchingStates: false,
 };